More changes in relation to Magnesium.Vulkan

+++ Callbacks/MgInternalAllocationType.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace Magnesium
 {
-    public enum MgInternalAllocationType : byte
+    public enum MgInternalAllocationType : uint
 	{
 		EXECUTABLE = 0,
 	};
